lift generateName() out of utils.ts
